APIS-5570 - Moving code around
